Crate oak_attestation_verification

Modules

expect

extract

policy

Contains code related to the policy that is combined from individual Event policies and is used to verify the whole TEE software stack.

verifier

Provides verification based on evidence, endorsements and reference values.

Traits

UnixTimestampMillis

Return a milliseconds-since-the-epoch timestamp value.

Functions

convert_pem_to_raw

Converts a PEM public key to raw ASN.1 DER bytes.

hex_to_raw_digest

Converts hex digest to raw digest.

raw_to_hex_digest

Converts raw digest to hex digest.

reference_values_from_evidence

verify_endorsement

Verifies a signed endorsement against a reference value.