Module oak_attestation_verification::claims

Contains structs for specifying in-toto statements and claims about software artifacts. See also https://github.com/project-oak/transparent-release/blob/main/pkg/claims/claim.go, https://github.com/project-oak/transparent-release/blob/main/pkg/intoto/intoto.go.

Structs

ClaimEvidence

Metadata about an artifact that serves as the evidence for the truth of a claim.

ClaimPredicate

Detailed content of a claim.

ClaimValidity

Validity time range of an issued claim.

Claimless

Inner type for a simple claim with no further fields.

Statement

Represents a generic statement that binds a predicate to a subject.

Subject

A software artifact identified by its name and a set of artifacts.

Enums

InvalidClaimData

Constants

CLAIM_V1

PredicateType which identifies a V1 Claim, for in-toto statements.

ENDORSEMENT_V2

ClaimType for endorsements. Expected to be used together with ClaimV1 as the predicate type in an in-toto statement.

STATEMENT_INTOTO_V01

URI representing in-toto v01 statements. This is constant for all predicate types.

Functions

get_digest

Assembles digests found in endorsement statement into a protocol buffer.

parse_endorsement_statement

Converts the given byte array into an endorsement statement.

validate_claim

Checks that the given statement is a valid claim:

validate_endorsement

Checks that the given endorsement statement is a valid and has the correct claim type.

verify_validity_duration

Checks that the input claim has a validity duration, and that the specified time is inside the validity period.

Type Aliases

DigestSet

EndorsementStatement