Trait oak_attestation_verification::policy::EventPolicy

pub trait EventPolicy {
    // Required method
    fn verify(
        &self,
        encoded_event: &[u8],
        encoded_event_endorsement: &[u8]
    ) -> Result<EventAttestationResults>;
}

Verification Policy that takes an encoded Event and an encoded Event Endorsement and performs attestation verification for this specific Event.

Required Methods

fn verify( &self, encoded_event: &[u8], encoded_event_endorsement: &[u8] ) -> Result<EventAttestationResults>