Module oak_attestation_verification::proto::oak::attestation::v1

Modules

attestation_results

Nested message and enum types in AttestationResults.

binary_reference_value

Nested message and enum types in BinaryReferenceValue.

endorsements

Nested message and enum types in Endorsements.

reference_values

Nested message and enum types in ReferenceValues.

Structs

AmdSevReferenceValues

ApplicationKeys

Keys used by the application to derive encryption session keys and to sign arbitrary data. Each of the certificates contains the final layer’s measurement as additional claims.

ApplicationLayerEndorsements

ApplicationLayerReferenceValues

Represents an application running under Oak Restricted Kernel.

AttestationResults

Represents a verification result. Can be extended to return certain measurements and other detail to the client for further processing. Nomenclature follows RFC 9334.

BinaryReferenceValue

CbEndorsements

CbLayerReferenceValues

CbReferenceValues

ContainerLayerEndorsements

ContainerLayerReferenceValues

Digests

Verifies that the field contains at least one of the given digests. No checks are performed if this is empty. A match in at least one digest is considered a success.

EndorsementReferenceValue

Verifies the transparency log entry, including signatures and the digest.

Endorsements

Endorsement contains statements that some entity (e.g. a hardware provider) vouches for the integrity of claims about the TEE or the software running on it.

Evidence

Attestation Evidence used by the client to the identity of firmware and software running inside a Trusted Execution Environment.

FileReferenceValue

Reference value for a file including its digests.

IntelTdxReferenceValues

TBD

KernelLayerEndorsements

KernelLayerReferenceValues

Reference values of the kernel layer, as measured by stage0.

LayerEvidence

DICE layer evidence containing a certificate signed by the previous layer.

OakContainersEndorsements

OakContainersReferenceValues

OakRestrictedKernelEndorsements

OakRestrictedKernelReferenceValues

ReferenceValues

RootLayerEndorsements

RootLayerEvidence

Evidence generated by the Layer0.

RootLayerReferenceValues

StringReferenceValue

Verifies that a particular string is equal to at least one of the specified ones. No checks are performed if this is empty.

SystemLayerEndorsements

SystemLayerReferenceValues

TransparentReleaseEndorsement

Endorsement for binaries from the Transparent Release process. <https://github.com/project-oak/oak/blob/main/docs/release.md>

Enums

TeePlatform

This proto defines the layered DICE Attestation Evidence.