Module oak_attestation_verification::rekor

This module provides structs for representing a Rekor LogEntry, as well as logic for parsing and verifying signatures in a Rekor LogEntry.

Structs

Body

Struct representing the body in a Rekor LogEntry.

Data

Struct representing the hashed data in the body of a Rekor LogEntry. Based on https://github.com/sigstore/rekor/blob/2978cdc26fdf8f5bfede8459afd9735f0f231a2a/pkg/generated/models/rekord_v001_schema.go#L179.

GenericSignature

Struct representing a signature in the body of a Rekor LogEntry. Based on https://github.com/sigstore/rekor/blob/2978cdc26fdf8f5bfede8459afd9735f0f231a2a/pkg/generated/models/rekord_v001_schema.go#L383

Hash

Struct representing a hash digest. Based on https://github.com/sigstore/rekor/blob/2978cdc26fdf8f5bfede8459afd9735f0f231a2a/pkg/generated/models/rekord_v001_schema.go#L273.

LogEntry

Struct representing a Rekor LogEntry. Based on https://github.com/sigstore/rekor/blob/2978cdc26fdf8f5bfede8459afd9735f0f231a2a/pkg/generated/models/log_entry.go#L89.

LogEntryVerification

Struct representing a verification object in a Rekor LogEntry. The verification object in Rekor also contains an inclusion proof. Since we currently don’t verify the inclusion proof in the client, it is omitted from this struct.

PublicKey

Struct representing a public key included in the body of a Rekor LogEntry. Based on https://github.com/sigstore/rekor/blob/2978cdc26fdf8f5bfede8459afd9735f0f231a2a/pkg/generated/models/rekord_v001_schema.go#L551.

RekorSignatureBundle

Convenient struct for verifying the signedEntryTimestamp in a Rekor LogEntry.

Spec

Struct representing the spec in the body of a Rekor LogEntry. Based on https://github.com/sigstore/rekor/blob/2978cdc26fdf8f5bfede8459afd9735f0f231a2a/pkg/generated/models/rekord_v001_schema.go#L39.

Functions

get_rekor_log_entry_body

Parses the given bytes into a Rekor LogEntry object, and returns its body parsed into an instance of Body.

verify_rekor_body

Verifies the signature in the body over the contents.

verify_rekor_log_entry

Verifies a Rekor LogEntry. This includes verifying:

verify_rekor_signature

Parses a blob into a Rekor log entry and verifies the signature in `signedEntryTimestamp`` using Rekor’s public key.