Module oak_attestation_verification::util

Enums

MatchResult

Functions

convert_pem_to_raw

Converts a PEM key to raw. Will panic if it does not look like PEM.

convert_pem_to_verifying_key

Converts a PEM-encoded x509/PKIX public key to a verifying key.

convert_raw_to_pem

Converts a raw public key to PEM format.

convert_raw_to_verifying_key

Converts a raw public key to a verifying key.

equal_keys

Compares two ECDSA public keys. Instead of comparing the bytes, we parse the bytes and compare p256 keys. Keys are considered equal if they are the same on the elliptic curve. This means that the keys could have different bytes, but still be the same key.

hash_sha2_256

hex_to_raw_digest

Converts hex digest to raw digest.

is_hex_digest_match

Compares two hex digests.

is_raw_digest_match

Compares two raw digests.

looks_like_pem

Makes a plausible guess whether the public key is in PEM format.

raw_digest_from_contents

Computes various digest formats of a binary array. The empty arrays need to be filled, when we depend on SHA{1,3} hashers.

raw_to_hex_digest

Converts raw digest to hex digest.

verify_signature_raw

Verifies the signature over the contents using the public key.