Module oak_attestation_verification::verifier

Provides verification based on evidence, endorsements and reference values.

Structs

AmdSevSnpDiceAttestationVerifier

Functions

get_event_artifact

Returns a reference to the event artifact from attestation_results with a given artifact_id.

to_attestation_results

verify

Verifies entire setup by forwarding to individual setup types.

verify_dice_chain

Verifies signatures of the certificates in the DICE chain and returns last layer’s Certificate Authority key if the verification is successful.

verify_dice_chain_and_extract_evidence

Verifies signatures of the certificates in the DICE chain and extracts the evidence values from the certificates if the verification is successful.

verify_with_expected_values

Verifies entire setup by forwarding to individual setup types.