Function oak_attestation_verification::verifier::verify

pub fn verify(
    now_utc_millis: i64,
    evidence: &Evidence,
    endorsements: &Endorsements,
    reference_values: &ReferenceValues,
) -> Result<ExtractedEvidence>

Verifies entire setup by forwarding to individual setup types.

This just fetches expected values using [expect::get_expected_values], and then call verify_with_expected_values providing those values.

If you’d like to cache and reuse the values, call those two methods independently, and cache the results of the first.