Crate oak_attestation_verification

Modules

amd

Contains code related to AMD hardware certificates and attestation report.

claims

Contains structs for specifying in-toto statements and claims about software artifacts. See also https://github.com/project-oak/transparent-release/blob/main/pkg/claims/claim.go, https://github.com/project-oak/transparent-release/blob/main/pkg/intoto/intoto.go.

endorsement

Verifies binary endorsements as coming from Transparent Release.

rekor

This module provides structs for representing a Rekor LogEntry, as well as logic for parsing and verifying signatures in a Rekor LogEntry.

util

verifier

Provides verification based on evidence, endorsements and reference values.