Module oak_attestation_verification::verifier

Provides verification based on evidence, endorsements and reference values.

Functions

extract_evidence

Extracts measurements, public keys and other attestation-related values from the evidence without verifying it. For most usecases, this function should not be used. Instead use the verify function, which verifies the attestation and only returns evidence upon successful verification. Hence marked as dangerous.

get_expected_values

to_attestation_results

verify

Verifies entire setup by forwarding to individual setup types. The now_utc_millis argument will be changed to a time type as work progresses.

verify_dice_chain

Verifies signatures of the certificates in the DICE chain and extracts the evidence values from the certificates if the verification is successful.

verify_with_expected_values

Verifies entire setup by forwarding to individual setup types. The now_utc_millis argument will be changed to a time type as work progresses. This variant returns expected values along with the extracted evidence.